Predict a bunch of stuff
