+++ Content.Server/Weapons/Melee/MeleeWeaponSystem.cs
@@ -1,5 +1,4 @@
 using System.Linq;
-using Content.Server.Actions.Events;
 using Content.Server.Body.Components;
 using Content.Server.Body.Systems;
 using Content.Server.Chemistry.Components;
@@ -8,6 +7,7 @@ using Content.Server.CombatMode.Disarm;
 using Content.Server.Contests;
 using Content.Server.Examine;
 using Content.Server.Movement.Systems;
+using Content.Shared.Actions.Events;
 using Content.Shared.Administration.Components;
 using Content.Shared.CombatMode;
 using Content.Shared.Damage;
