+++ Content.Shared/Wieldable/BeforeWieldEvent.cs
@@ -1,0 +1,5 @@
+namespace Content.Shared.Wieldable;
+
+public sealed class BeforeWieldEvent : CancellableEntityEventArgs
+{
+}
